Refactor legacy Grasshopper assets into components and kernel
